Moved modelstate to different class
